backend/pyproject.toml
[tool.poetry]
name = "own-prj"
version = "0.1.0"
description = ""
authors = ["Bezizvestnyi <chikovm@bk.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
psycopg2 = "^2.9.10"
django-environ = "^0.12.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
djoser = "^2.3.1"
djangorestframework-simplejwt = "^5.5.0"
drf-yasg = "^1.21.9"
social-auth-app-django = "^5.4.3"
pillow = "^11.1.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
